Improve FSDP support (SimCLR + CAPI) and fix PyTorch 2.10 custom CUDA memory leak
Added:
  - Add VOLO (Vision Outlooker) classification models with outlook + class attention
  - Add cosine_fraction to cosine_scheduler/get_scheduler for truncated cosine annealing

Changed:
  - Improve reliability of train_simclr --distributed-mode fsdp (beta)
  - Add train_capi --distributed-mode fsdp with two roots (student + teacher)
  - Extend fsdp_utils.setup_fsdp to accept optional prebuilt device mesh for mesh reuse

Fixed:
  - Fix PyTorch 2.10 custom CUDA op memory leak (MSDA/SWAttention) by migrating bindings to TORCH_LIBRARY_EXPAND
  - Optimize MSDA pure-Python path by removing GPU→CPU sync and reducing allocations
  - Fix layer-decay + custom weight-decay grouping for torch.compile-wrapped models
  - Keep Data2Vec/Data2Vec2 EMA teacher backbone in eval mode during training

+++ birder/net/ssl/capi.py
@@ -331,7 +331,7 @@ class Decoder(nn.Module):
     def mask_tokens_grid(self, mask: torch.Tensor) -> torch.Tensor:
         N = mask.size(0)
         x = self.decoder_pos_embed.masked_select(mask.bool().unsqueeze(-1)).reshape(N, -1, self.mask_token.size(-1))
-        x = x + self.mask_token
+        x = x.to(self.mask_token.dtype) + self.mask_token
 
         return x
